requirements.txt
asgiref==3.4.1
black==21.11b1
click==8.0.3
Django==3.2.9
django-htmx==1.6.0
django-localflavor==3.1
django-widget-tweaks==1.4.9
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-stdnum==1.17
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.1
